Try to create the ramdisk on mountpoint only if it doesn't already exists.

+++ kill-ryzen.sh
@@ -5,6 +5,7 @@ USE_RAMDISK=true
 CLEAN_ON_EXIT=false
 NPROC=$1
 TPROC=$2
+MOUNTPOINT="/mnt/ramdisk"
 
 [ -n "$NPROC" ] || NPROC=$(nproc)
 [ -n "$TPROC" ] || TPROC=1
@@ -30,16 +31,22 @@ else
 fi
 
 if $USE_RAMDISK; then
-  echo "Create compressed ramdisk"
-  sudo mkdir -p /mnt/ramdisk || exit 1
-  sudo modprobe zram num_devices=1 || exit 1
-  echo 64G | sudo tee /sys/block/zram0/disksize || exit 1
-  sudo mkfs.ext4 -q -m 0 -b 4096 -O sparse_super -L zram /dev/zram0 || exit 1
-  sudo mount -o relatime,nosuid,discard /dev/zram0 /mnt/ramdisk/ || exit 1
-  sudo mkdir -p /mnt/ramdisk/workdir || exit 1
-  sudo chmod 777 /mnt/ramdisk/workdir || exit 1
-  cp buildloop.sh /mnt/ramdisk/workdir/buildloop.sh || exit 1
-  cd /mnt/ramdisk/workdir || exit 1
+  #Create compressed ramdisk if it doesn't already exists
+  if mountpoint -q $MOUNTPOINT; then
+    echo "RAMdisk already mounted on $MOUNTPOINT."
+  else
+    echo "Create compressed ramdisk"
+    sudo mkdir -p $MOUNTPOINT || exit 1
+    sudo modprobe zram num_devices=1 || exit 1
+    # Setting maximum RAM disk size
+    echo 64G | sudo tee /sys/block/zram0/disksize || exit 1
+    sudo mkfs.ext4 -q -m 0 -b 4096 -O sparse_super -L zram /dev/zram0 || exit 1
+    sudo mount -o relatime,nosuid,discard /dev/zram0 $MOUNTPOINT || exit 1
+  fi
+  sudo mkdir -p $MOUNTPOINT/workdir || exit 1
+  sudo chmod 777 $MOUNTPOINT/workdir || exit 1
+  cp buildloop.sh $MOUNTPOINT/workdir/buildloop.sh || exit 1
+  cd $MOUNTPOINT/workdir || exit 1
   mkdir -p tmpdir || exit 1
   export TMPDIR="$PWD/tmpdir"
 fi
